refactor (whisper): move whisper services into dedicated folder and split WhisperService by responsibility
- Move Whisper files under Core/Services/Whisper
- Split WhisperService into base, model lifecycle, and transcription core extension files
- Keep behavior and external service interface unchanged
- Update Xcode project groups/source membership for moved and new files
- Update CODEMAP and ENGINEERING docs to reflect new Whisper structure
