+++ Core/Services/Whisper/WhisperService+TranscriptionCore.swift
@@ -1,88 +1,8 @@
 import Foundation
-import KeyVoxWhisper
-import Combine
 import AVFoundation
+import KeyVoxWhisper
 
-class WhisperService: ObservableObject {
-    @Published var isTranscribing = false
-    @Published var transcriptionText = ""
-    @Published private(set) var lastResultWasLikelyNoSpeech = false
-    
-    private var whisper: Whisper?
-    private var dictionaryHintPrompt = ""
-    private let noSpeechSegmentProbabilityThreshold: Float = 0.72
-    private let noSpeechAverageProbabilityThreshold: Float = 0.80
-    private let paragraphChunker = WhisperAudioParagraphChunker()
-    // Enabled by default; temporarily disable locally when validating phonetic matching without hint bias.
-    private let isPromptHintingEnabled = true
-    private let suspiciousShortResultMinChunkSeconds: Double = 1.35
-    private let suspiciousShortResultMaxWords = 2
-    private let suspiciousShortResultMaxNoSpeechProbability: Float = 0.35
-    private let retryRelaxedLogprobThreshold: Float = -2.0
-    
-    /// Pre-loads the model into memory to eliminate cold-start latency.
-    func warmup() {
-        guard whisper == nil else {
-            #if DEBUG
-            print("WhisperService: Warmup skipped (model already loaded).")
-            #endif
-            return
-        }
-        guard let modelPath = getModelPath() else {
-            #if DEBUG
-            print("WhisperService: Warmup skipped (model files not found).")
-            #endif
-            return
-        }
-        
-        #if DEBUG
-        print("Warming up Whisper model with optimized settings...")
-        #endif
-        
-        let params = WhisperParams.default
-        params.language = .auto
-        params.n_threads = 4 // Optimal for M-series P-cores (prevent oversubscription)
-        params.no_context = true
-        params.print_timestamps = false
-        params.suppress_blank = true
-        params.suppress_non_speech_tokens = true
-        params.temperature = 0.0
-        params.temperature_inc = 0.0
-        params.no_speech_thold = 0.6
-        params.logprob_thold = -0.8
-        params.initialPrompt = isPromptHintingEnabled ? dictionaryHintPrompt : ""
-        // CoreML is automatic if the model files are present
-        
-        whisper = Whisper(fromFileURL: URL(fileURLWithPath: modelPath), withParams: params)
-    }
-
-    /// Unloads the currently cached model instance.
-    /// Used when model files are deleted so re-download can warm from disk again.
-    func unloadModel() {
-        guard whisper != nil else { return }
-        whisper = nil
-        #if DEBUG
-        print("WhisperService: Unloaded model from memory.")
-        #endif
-    }
-    
-    private var transcriptionTask: Task<Void, Never>?
-    
-    func cancelTranscription() {
-        transcriptionTask?.cancel()
-        transcriptionTask = nil
-        #if DEBUG
-        print("WhisperService: Transcription cancelled.")
-        #endif
-    }
-
-    func updateDictionaryHintPrompt(_ prompt: String) {
-        let cleanedPrompt = prompt
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        dictionaryHintPrompt = cleanedPrompt
-        whisper?.params.initialPrompt = isPromptHintingEnabled ? cleanedPrompt : ""
-    }
-    
+extension WhisperService {
     func transcribe(
         audioFrames: [Float],
         useDictionaryHintPrompt: Bool = true,
@@ -99,10 +19,10 @@ class WhisperService: ObservableObject {
             completion("")
             return
         }
-        
+
         self.isTranscribing = true
         self.lastResultWasLikelyNoSpeech = false
-        
+
         // Ensure model is loaded (if warmup wasn't called/finished)
         if whisper == nil {
             warmup()
@@ -123,11 +43,11 @@ class WhisperService: ObservableObject {
             guard let self, !shouldUseDictionaryHintPrompt else { return }
             self.whisper?.params.initialPrompt = self.isPromptHintingEnabled ? self.dictionaryHintPrompt : ""
         }
-        
+
         #if DEBUG
         print("Transcribing \(audioFrames.count) raw frames...")
         #endif
-        
+
         transcriptionTask = Task {
             do {
                 let chunkResult = self.paragraphChunker.split(audioFrames)
@@ -169,7 +89,7 @@ class WhisperService: ObservableObject {
                         chunkTexts.append(normalizedChunkText)
                     }
                 }
-                
+
                 // Check if task was cancelled before proceeding
                 if Task.isCancelled {
                     DispatchQueue.main.async {
@@ -178,7 +98,7 @@ class WhisperService: ObservableObject {
                     }
                     return
                 }
-                
+
                 let paragraphSeparator = enableAutoParagraphs ? "\n\n" : " "
                 let text = chunkTexts
                     .joined(separator: paragraphSeparator)
@@ -219,7 +139,7 @@ class WhisperService: ObservableObject {
                     }
                     return
                 }
-                
+
                 #if DEBUG
                 print("Transcription error: \(error)")
                 #endif
@@ -232,7 +152,7 @@ class WhisperService: ObservableObject {
             }
         }
     }
-    
+
     // Legacy support for file-based transcription (can be removed later)
     func transcribe(audioURL: URL, completion: @escaping (String?) -> Void) {
         Task {
@@ -244,24 +164,24 @@ class WhisperService: ObservableObject {
             }
         }
     }
-    
+
     private func loadAndResample(url: URL) throws -> [Float] {
         let inputFile = try AVAudioFile(forReading: url)
         let inputFormat = inputFile.processingFormat
-        
+
         let outputFormat = AVAudioFormat(commonFormat: .pcmFormatFloat32, sampleRate: 16000, channels: 1, interleaved: false)!
-        
+
         guard let converter = AVAudioConverter(from: inputFormat, to: outputFormat) else {
             throw NSError(domain: "WhisperService", code: 1, userInfo: [NSLocalizedDescriptionKey: "Failed to create converter"])
         }
-        
+
         let ratio = 16000.0 / inputFormat.sampleRate
         let outputCapacity = AVAudioFrameCount(Double(inputFile.length) * ratio) + 1
-        
+
         guard let outputBuffer = AVAudioPCMBuffer(pcmFormat: outputFormat, frameCapacity: outputCapacity) else {
             throw NSError(domain: "WhisperService", code: 2, userInfo: [NSLocalizedDescriptionKey: "Failed to create buffer"])
         }
-        
+
         var error: NSError?
         let inputBlock: AVAudioConverterInputBlock = { inNumPackets, outStatus in
             let inputBuffer = AVAudioPCMBuffer(pcmFormat: inputFormat, frameCapacity: inNumPackets)!
@@ -274,29 +194,15 @@ class WhisperService: ObservableObject {
                 return nil
             }
         }
-        
+
         let status = converter.convert(to: outputBuffer, error: &error, withInputFrom: inputBlock)
-        
+
         if status == .error {
             throw error ?? NSError(domain: "WhisperService", code: 3, userInfo: [NSLocalizedDescriptionKey: "Conversion failed"])
         }
-        
+
         guard let floatData = outputBuffer.floatChannelData else { return [] }
         return Array(UnsafeBufferPointer(start: floatData[0], count: Int(outputBuffer.frameLength)))
-    }
-    
-    private func getModelPath() -> String? {
-        let appSupport = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!
-        let modelsDir = appSupport
-            .appendingPathComponent("KeyVox")
-            .appendingPathComponent("Models")
-        
-        let baseModelURL = modelsDir.appendingPathComponent("ggml-base.bin")
-        if FileManager.default.fileExists(atPath: baseModelURL.path) {
-            return baseModelURL.path
-        }
-        
-        return nil
     }
 
     private func transcribeChunkWithLeadingPhraseRetry(
