+++ Docs/CODEMAP.md
@@ -54,8 +54,11 @@ KeyVox/
 │   │   │   └── PasteSpacingHeuristics.swift
 │   │   ├── UpdatePromptPresenting.swift
 │   │   ├── UpdateFeedConfig.swift
-│   │   ├── WhisperAudioParagraphChunker.swift
-│   │   └── WhisperService.swift
+│   │   └── Whisper/
+│   │       ├── WhisperAudioParagraphChunker.swift
+│   │       ├── WhisperService.swift
+│   │       ├── WhisperService+ModelLifecycle.swift
+│   │       └── WhisperService+TranscriptionCore.swift
 │   ├── Language/
 │   │   ├── Dictionary/
 │   │   │   ├── Email/
@@ -244,8 +247,8 @@ KeyVox/
 1. `Core/KeyboardMonitor.swift` publishes trigger/shift/escape/caps-lock state.
 2. `Core/Transcription/TranscriptionManager.swift` drives app state: `idle -> recording -> transcribing -> idle`.
 3. `Core/Audio/AudioRecorder.swift` captures live audio as mono float frames at 16kHz.
-4. `Core/Services/WhisperAudioParagraphChunker.swift` detects long internal silence and computes conservative chunk boundaries.
-5. `Core/Services/WhisperService.swift` transcribes each chunk through `KeyVoxWhisper` and stitches chunks with paragraph or space separators.
+4. `Core/Services/Whisper/WhisperAudioParagraphChunker.swift` detects long internal silence and computes conservative chunk boundaries.
+5. `Core/Services/Whisper/WhisperService.swift` transcribes each chunk through `KeyVoxWhisper` and stitches chunks with paragraph or space separators.
 6. `Core/Transcription/TranscriptionPostProcessor.swift` orchestrates dictionary correction, list formatting, and specialized normalization helpers under `Core/Normalization/`.
 7. `Core/Services/Paste/PasteService.swift` inserts text via Accessibility first, then menu-bar Paste fallback.
 8. `Core/Overlay/OverlayManager.swift` owns overlay lifecycle orchestration and delegates motion/persistence helpers.
@@ -350,13 +353,17 @@ KeyVox/
 
 ### Service Layer (`Core/Services`)
 
-- `Core/Services/WhisperAudioParagraphChunker.swift`
+- `Core/Services/Whisper/WhisperAudioParagraphChunker.swift`
   - Splits long captures into paragraph-sized chunks using deterministic RMS silence windows.
   - Uses configurable chunk-size and silence-run guardrails to avoid over-splitting.
-- `Core/Services/WhisperService.swift`
+- `Core/Services/Whisper/WhisperService.swift`
   - Loads model from Application Support and runs inference.
   - Uses automatic language detection (`.auto`).
   - Supports optional auto-paragraph stitching via `enableAutoParagraphs`.
+- `Core/Services/Whisper/WhisperService+ModelLifecycle.swift`
+  - Isolates model lifecycle helpers (`warmup`, `unloadModel`, model-path resolution).
+- `Core/Services/Whisper/WhisperService+TranscriptionCore.swift`
+  - Owns chunk transcription flow, retry selection, whitespace normalization, and debug segment logging.
 
 ### Post-Processing (`Core` + `Core/Normalization` + `Core/Language` + `Core/Lists`)
